Extract tb fetching for ECDR initial processing to `tb_data` module
