+++ seaice_ecdr/initial_daily_ecdr.py
@@ -10,7 +10,7 @@ import sys
 import traceback
 from functools import cache
 from pathlib import Path
-from typing import Final, Iterable, TypedDict, cast, get_args
+from typing import Iterable, TypedDict, cast, get_args
 
 import click
 import numpy as np
@@ -34,10 +34,7 @@ from pm_icecon.land_spillover import apply_nt2_land_spillover
 from pm_icecon.nt._types import NasateamGradientRatioThresholds
 from pm_icecon.nt.tiepoints import NasateamTiePoints
 from pm_tb_data._types import NORTH, Hemisphere
-from pm_tb_data.fetch.amsr.ae_si import get_ae_si_tbs_from_disk
-from pm_tb_data.fetch.amsr.au_si import get_au_si_tbs
-from pm_tb_data.fetch.amsr.util import AMSR_RESOLUTIONS
-from pm_tb_data.fetch.nsidc_0001 import NSIDC_0001_SATS, get_nsidc_0001_tbs_from_disk
+from pm_tb_data.fetch.nsidc_0001 import NSIDC_0001_SATS
 
 from seaice_ecdr._types import ECDR_SUPPORTED_RESOLUTIONS
 from seaice_ecdr.ancillary import (
@@ -53,6 +50,7 @@ from seaice_ecdr.grid_id import get_grid_id
 from seaice_ecdr.gridid_to_xr_dataarray import get_dataset_for_grid_id
 from seaice_ecdr.platforms import get_platform_by_date
 from seaice_ecdr.regrid_25to12 import reproject_ideds_25to12
+from seaice_ecdr.tb_data import get_ecdr_tbs
 from seaice_ecdr.util import date_range, standard_daily_filename
 
 EXPECTED_TB_NAMES = ("h18", "v18", "v23", "h36", "v36")
@@ -257,21 +255,6 @@ def _setup_ecdr_ds(
         )
 
     return ecdr_ide_ds
-
-
-def _pm_icecon_amsr_res_str(
-    *, resolution: ECDR_SUPPORTED_RESOLUTIONS
-) -> AMSR_RESOLUTIONS:
-    """Given an AMSR ECDR resolution string, return a compatible `pm_icecon` resolution string."""
-    _ecdr_pm_icecon_resolution_mapping: dict[
-        ECDR_SUPPORTED_RESOLUTIONS, AMSR_RESOLUTIONS
-    ] = {
-        "12.5": "12",
-        "25": "25",
-    }
-    au_si_resolution_str = _ecdr_pm_icecon_resolution_mapping[resolution]
-
-    return au_si_resolution_str
 
 
 def compute_initial_daily_ecdr_dataset(
@@ -809,129 +792,6 @@ def compute_initial_daily_ecdr_dataset(
     return ecdr_ide_ds
 
 
-# TODO: Eventually, these should be renamed based on microwave
-#       band names, not AMSR2 TB channel frequencies
-def rename_0001_tbs(
-    *,
-    input_ds: xr.Dataset,
-) -> xr.Dataset:
-    """Rename 0001 TB fields for use with siconc code written for AMSR2."""
-    nsidc0001_to_amsr_tb_mapping = {
-        "h19": "h18",
-        "v19": "v18",
-        "v22": "v23",
-        "h37": "h36",
-        "v37": "v36",
-    }
-
-    new_tbs = {}
-    for key in input_ds.data_vars.keys():
-        data_var = input_ds.data_vars[key]
-        new_key = nsidc0001_to_amsr_tb_mapping[key]
-        new_tbs[new_key] = xr.DataArray(
-            data_var.data,
-            dims=("fake_y", "fake_x"),
-            attrs=data_var.attrs,
-        )
-
-    new_ds = xr.Dataset(new_tbs)
-
-    return new_ds
-
-
-def create_null_au_si_tbs(
-    *,
-    hemisphere: Hemisphere,
-    resolution: ECDR_SUPPORTED_RESOLUTIONS,
-) -> xr.Dataset:
-    """Create xr dataset containing all null-value data for all tbs."""
-    chan_desc = {
-        "h18": "18.7 GHz horizontal daily average Tbs",
-        "v18": "18.7 GHz vertical daily average Tbs",
-        "h23": "23.8 GHz horizontal daily average Tbs",
-        "v23": "23.8 GHz vertical daily average Tbs",
-        "h36": "36.5 GHz horizontal daily average Tbs",
-        "v36": "36.5 GHz vertical daily average Tbs",
-        "h89": "89.0 GHz horizontal daily average Tbs",
-        "v89": "89.0 GHz vertical daily average Tbs",
-    }
-
-    if hemisphere == "north" and resolution == "12.5":
-        xdim = 608
-        ydim = 896
-    elif hemisphere == "south" and resolution == "12.5":
-        xdim = 632
-        ydim = 664
-    else:
-        value_error_string = f"""
-        Could not create null_set of TBs for
-          hemisphere: {hemisphere}
-          resolution: {resolution}
-        """
-        raise ValueError(value_error_string)
-
-    null_array = np.zeros((ydim, xdim), dtype=np.float64)
-    null_array[:] = np.nan
-    common_tb_attrs = {
-        "_FillValue": 0,
-        "units": "degree_kelvin",
-        "standard_name": "brightness_temperature",
-        "packing_convention": "netCDF",
-        "packing_convention_description": "unpacked = scale_factor x packed + add_offset",
-        "scale_factor": 0.1,
-        "add_offset": 0.0,
-    }
-    null_tbs = xr.Dataset(
-        data_vars=dict(
-            h18=(
-                ["YDim", "XDim"],
-                null_array,
-                common_tb_attrs,
-            ),
-            v18=(
-                ["YDim", "XDim"],
-                null_array,
-                common_tb_attrs,
-            ),
-            h23=(
-                ["YDim", "XDim"],
-                null_array,
-                common_tb_attrs,
-            ),
-            v23=(
-                ["YDim", "XDim"],
-                null_array,
-                common_tb_attrs,
-            ),
-            h36=(
-                ["YDim", "XDim"],
-                null_array,
-                common_tb_attrs,
-            ),
-            v36=(
-                ["YDim", "XDim"],
-                null_array,
-                common_tb_attrs,
-            ),
-            h89=(
-                ["YDim", "XDim"],
-                null_array,
-                common_tb_attrs,
-            ),
-            v89=(
-                ["YDim", "XDim"],
-                null_array,
-                common_tb_attrs,
-            ),
-        ),
-    )
-
-    for key in chan_desc.keys():
-        null_tbs.data_vars[key].attrs.update({"long_name": chan_desc[key]})
-
-    return null_tbs
-
-
 def reproject_ideds(
     initial_ecdr_ds: xr.Dataset,
     date: dt.date,
@@ -964,87 +824,18 @@ def initial_daily_ecdr_dataset(
     hemisphere: Hemisphere,
     resolution: ECDR_SUPPORTED_RESOLUTIONS,
 ) -> xr.Dataset:
-    """Create xr dataset containing the first pass of daily enhanced CDR.
-
-    This uses AU_SI12 TBs
-    and others..."""
-    platform = get_platform_by_date(date)
-    if platform == "am2":
-        au_si_resolution_str = _pm_icecon_amsr_res_str(resolution=resolution)
-        try:
-            xr_tbs = get_au_si_tbs(
-                date=date,
-                hemisphere=hemisphere,
-                resolution=au_si_resolution_str,
-            )
-        except FileNotFoundError:
-            xr_tbs = create_null_au_si_tbs(
-                hemisphere=hemisphere,
-                resolution=resolution,
-            )
-            logger.warning(
-                f"Used all-null TBS for date={date},"
-                f" hemisphere={hemisphere},"
-                f" resolution={resolution}"
-            )
-        tb_resolution = resolution
-    elif platform == "ame":
-        ame_resolution_str = _pm_icecon_amsr_res_str(resolution=resolution)
-        AME_DATA_DIR = Path("/ecs/DP4/AMSA/AE_SI12.003/")
-        try:
-            xr_tbs = get_ae_si_tbs_from_disk(
-                date=date,
-                hemisphere=hemisphere,
-                data_dir=AME_DATA_DIR,
-                resolution=ame_resolution_str,
-            )
-        except FileNotFoundError:
-            logger.warning(f"Using null AU_SI12 for AME on {date}")
-            xr_tbs = create_null_au_si_tbs(
-                hemisphere=hemisphere,
-                resolution=resolution,
-            )
-            logger.warning(
-                f"Used all-null TBS for date={date},"
-                f" hemisphere={hemisphere},"
-                f" resolution={resolution}"
-            )
-        tb_resolution = resolution
-    elif platform == "F17":
-        NSIDC0001_DATA_DIR = Path("/ecs/DP4/PM/NSIDC-0001.006/")
-        # NSIDC-0001 TBs for siconc are all at 25km
-        nsidc0001_resolution: Final = "25"
-        try:
-            xr_tbs_0001 = get_nsidc_0001_tbs_from_disk(
-                date=date,
-                hemisphere=hemisphere,
-                data_dir=NSIDC0001_DATA_DIR,
-                resolution=nsidc0001_resolution,
-                sat=platform,
-            )
-            xr_tbs = rename_0001_tbs(input_ds=xr_tbs_0001)
-        except FileNotFoundError:
-            logger.warning(f"Using null TBs for {platform} on {date}")
-            print("this needs to be create_null_0001_tbs()")
-            xr_tbs = create_null_au_si_tbs(
-                hemisphere=hemisphere,
-                resolution=resolution,
-            )
-            logger.warning(
-                f"Used all-null TBS for date={date},"
-                f" hemisphere={hemisphere},"
-                f" resolution={resolution}"
-            )
-        tb_resolution = nsidc0001_resolution
-    else:
-        raise RuntimeError(f"Platform not supported: {platform}")
-
+    """Create xr dataset containing the first pass of daily enhanced CDR."""
+    tb_data = get_ecdr_tbs(
+        date=date,
+        hemisphere=hemisphere,
+        resolution=resolution,
+    )
     initial_ecdr_ds = compute_initial_daily_ecdr_dataset(
         date=date,
         hemisphere=hemisphere,
         resolution=resolution,
-        xr_tbs=xr_tbs,
-        tb_resolution=tb_resolution,
+        xr_tbs=tb_data.tbs,
+        tb_resolution=tb_data.resolution,
     )
 
     # If the computed ide_ds is not on the desired grid (ie resolution),
@@ -1052,12 +843,12 @@ def initial_daily_ecdr_dataset(
     # In general, this should be a comparison of grid_id's, but for
     # eCDR, a comparison of resolutions will suffice
     # Finished!
-    if resolution != tb_resolution:
+    if resolution != tb_data.resolution:
         initial_ecdr_ds = reproject_ideds(
             initial_ecdr_ds=initial_ecdr_ds,
             date=date,
             hemisphere=hemisphere,
-            tb_resolution=tb_resolution,
+            tb_resolution=tb_data.resolution,
             resolution=resolution,
         )
         logger.info(f"Reprojected ide_ds to {resolution}km")
